+++ modules/backend/behaviors/RelationController.php
@@ -414,7 +414,7 @@ class RelationController extends ControllerBehavior
      */
     protected function findExistingRelationIds($checkIds = null)
     {
-        $foreignKeyName = $this->relationModel->getKeyName();
+        $foreignKeyName = $this->relationModel->getQualifiedKeyName();
 
         $results = $this->relationObject
             ->getBaseQuery()